requirements.txt
discord.py[voice]>=2.3.0
Pillow>=10.0.0
